Spatial hdf5

+++ vcell-cli/src/main/java/org/vcell/cli/run/ExecuteImpl.java
@@ -176,7 +176,7 @@ public class ExecuteImpl {
             logger.info("Initializing SED-ML document...");
             String sedmlName = "", logDocumentMessage = "Initializing SED-ML document... ", logDocumentError = "";
             boolean somethingFailed = false; // shows that the current document suffered a partial or total failure
-            Map<TaskJob, ODESolverResultSet> resultsHash;
+
             File outDirForCurrentSedml = new File(omexHandler.getOutputPathFromSedml(sedmlLocation));
 
             try {
@@ -281,14 +281,12 @@ public class ExecuteImpl {
              */
             SolverHandler solverHandler = new SolverHandler();
             ExternalDocInfo externalDocInfo = new ExternalDocInfo(new File(inputFilePath), true);
-            resultsHash = new LinkedHashMap<TaskJob, ODESolverResultSet>();
             try {
                 String str = "Building solvers and starting simulation of all tasks... ";
                 logger.info(str);
                 logDocumentMessage += str;
-                resultsHash = solverHandler.simulateAllTasks(externalDocInfo, sedml, cliLogger, outDirForCurrentSedml, outputDir,
+                solverHandler.simulateAllTasks(externalDocInfo, sedml, cliLogger, outDirForCurrentSedml, outputDir,
                         outputBaseDir, sedmlLocation, bKeepTempFiles, bExactMatchOnly);
-                //Map<String, String> sim2Hdf5Map = solverHandler.sim2Hdf5Map;    // may not need it
             } catch (Exception e) {
                 somethingFailed = true;
                 anySedmlDocumentHasFailed = true;
@@ -317,47 +315,58 @@ public class ExecuteImpl {
             // that will include spatial simulations for which we don't produce reports or plots!
             //
             try {
-                if (resultsHash.containsValue(null)) {        // some tasks failed, but not all
+                if (solverHandler.nonSpatialResults.containsValue(null) || solverHandler.spatialResults.containsValue(null)) {        // some tasks failed, but not all
                     anySedmlDocumentHasFailed = true;
                     somethingFailed = true;
                     logDocumentMessage += "Failed to execute one or more tasks. ";
                     logger.info("Failed to execute one or more tasks in " + sedmlName);
                 }
 
-                HashMap<String, File> reportsHash = RunUtils.generateReportsAsCSV(sedml, resultsHash, outDirForCurrentSedml, outputDir, sedmlLocation);
-                if (reportsHash == null || reportsHash.size() == 0) {
-                    anySedmlDocumentHasFailed = true;
-                    somethingFailed = true;
-                    String msg = "Failed to generate any reports. ";
-                    throw new RuntimeException(msg);
-                }
-                if (reportsHash.containsValue(null)) {
-                    anySedmlDocumentHasFailed = true;
-                    somethingFailed = true;
-                    String msg = "Failed to generate one or more reports. ";
-                    logDocumentMessage += msg;
-                } else {
-                    logDocumentMessage += "Done. ";
-                }
-                String idNamePlotsMap = RunUtils.generateIdNamePlotsMap(sedml, outDirForCurrentSedml);
-                PythonCalls.execPlotOutputSedDoc(inputFilePath, idNamePlotsMap, outputDir);                            // create the HDF5 file
+                logDocumentMessage += "Generating outputs... ";
+                logger.info("Generating outputs... ");
 
-                logger.info("Generating Plots... ");
-                PythonCalls.genPlotsPseudoSedml(sedmlLocation, outDirForCurrentSedml.toString());    // generate the plots
+                if (!solverHandler.nonSpatialResults.isEmpty()) {
+	                HashMap<String, File> csvReports = null;
+	                logDocumentMessage += "Generating CSV file... ";
+	                logger.info("Generating CSV file... ");
+                  
+	                csvReports = RunUtils.generateReportsAsCSV(sedml, solverHandler.nonSpatialResults, outDirForCurrentSedml, outputDir, sedmlLocation);
+	                String idNamePlotsMap = RunUtils.generateIdNamePlotsMap(sedml, outDirForCurrentSedml);
+	                PythonCalls.execPlotOutputSedDoc(inputFilePath, idNamePlotsMap, outputDir);                            // create the HDF5 file
+	
+	                if (csvReports == null || csvReports.isEmpty() || csvReports.containsValue(null)) {
+	                    anySedmlDocumentHasFailed = true;
+	                    somethingFailed = true;
+	                    String msg = "Failed to generate one or more reports. ";
+	                    logDocumentMessage += msg;
+	                } else {
+	                    logDocumentMessage += "Done. ";
+	                }
+                  
+                  logger.info("Generating Plots... ");
+                  PythonCalls.genPlotsPseudoSedml(sedmlLocation, outDirForCurrentSedml.toString());    // generate the plots
 
-                // remove CSV files associated with reports, these values are in report.h5 file anyway
-                for (File file : reportsHash.values()){
-                    file.delete();
+                  // remove CSV files associated with reports, these values are in report.h5 file anyway
+                  for (File file : csvReports.values()){
+                      file.delete();
+                  }
+                  logDocumentMessage += "Generating HDF5 file... ";
+                  logger.info("Generating HDF5 file... ");
+                  RunUtils.generateReportsAsHDF5(sedml, solverHandler.nonSpatialResults, outDirForCurrentSedml, sedmlLocation);	
+	
+	                if (!containsExtension(outputDir, "h5")) {
+	                    anySedmlDocumentHasFailed = true;
+	                    somethingFailed = true;
+	                    throw new RuntimeException("Failed to generate the HDF5 output file. ");
+	                } else {
+	                    logDocumentMessage += "Done. ";
+	                }
                 }
-                logDocumentMessage += "Generating HDF5 file... ";
-                logger.info("Generating HDF5 file... ");
-                RunUtils.generateReportsAsHDF5(sedml, resultsHash, outDirForCurrentSedml, sedmlLocation);
-                if (!containsExtension(outDirForCurrentSedml.getAbsolutePath(), "h5")) {
-                    anySedmlDocumentHasFailed = true;
-                    somethingFailed = true;
-                    throw new RuntimeException("Failed to generate the HDF5 output file. ");
-                } else {
-                    logDocumentMessage += "Done. ";
+                
+                if (!solverHandler.spatialResults.isEmpty()) {
+                	// TODO
+                	// check for failures
+                	// generate reports from hdf5 outputs and add to non-spatial reports, if any
                 }
